+++ Docs/CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.690] - 2025-10-20
+## [0.7.750] - 2025-10-20
 ### Fixed
 - Fixing proxy creation in `OpenTgResearcherConsole`
 - Fixing the `Client Connection` page in `OpenTgResearcherDesktop`
 - Fixing the creation of the `Saved Messages` chat in `OpenTgResearcherDesktop`
+- Fixing storage initialization in `OpenTgResearcherDesktop`
+- Fixing language change in `OpenTgResearcherDesktop`
 ### Changed
 - Updating the `Update` page in `OpenTgResearcherDesktop`
 ### Added